+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/blocks/schema-config-admin-fields/src/index.js
@@ -30,7 +30,7 @@ registerBlockType( 'graphql-api/schema-config-admin-fields', {
 	 * This is the display title for your block, which can be translated with `i18n` functions.
 	 * The block inserter will show this name.
 	 */
-	title: __( '"Schema Admin Fields" for the Schema Configuration', 'graphql-api' ),
+	title: __( '"Schema Expose Admin Data" for the Schema Configuration', 'graphql-api' ),
 
 	/**
 	 * This is a short description for your block, can be translated with `i18n` functions.